Fixing bin file path evaluation; extracting code block to a new utility
